requirements.txt
# Interfaz y Dashboard
streamlit==1.32.0
plotly==5.19.0
openpyxl==3.1.2

# Ciencia de Datos y Machine Learning
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.1.post1
xgboost==2.0.3
joblib==1.3.2

# Inteligencia Artificial y RAG
langchain==0.1.11
langchain-community==0.0.27
langchain-groq==0.0.1
langchain-core==0.1.30
chromadb==0.4.24
sentence-transformers==2.5.1
rank-bm25==0.2.2
groq==0.4.2

# Utilidades
python-dotenv==1.0.1